project/broadcaster/requirements.txt
requests
nats-py